refactor: move API-handling function to separate folder
