+++ src/js/ApiGetter/ApiGetter.js
@@ -1,12 +1,7 @@
 import React, { Component } from 'react';
 import '../../scss/main.scss';
 
-const API_KEY = 'LR7eqPMCqaBKMlIEKf17ExpE4DlGwExDHLCXhGrZ';
-
-const TEST_ROVER = 'curiosity';
-const TEST_SOL = 1000;
-const TEST_CAMERA = 'fhaz';
-const TEST_PAGE = 1;
+import getInfoFromAPI from '../api/api';
 
 class ApiGetter extends Component {
   constructor(props) {
@@ -15,27 +10,6 @@ class ApiGetter extends Component {
       table: [],
     };
   }
-
-  getInfoFromAPI = async (rover = TEST_ROVER, sol = TEST_SOL, camera = TEST_CAMERA, page = TEST_PAGE) => {
-    const tableInfo = [];
-    const url = `https://api.nasa.gov/mars-photos/api/v1/rovers/${rover}/photos?sol=${sol}&camera=${camera}&page=${page}&api_key=${API_KEY}`;
-    const response = await fetch(url);
-    const responseJson = await response.json();
-    await responseJson.photos.forEach(elem => {
-      const tempDay = {
-        imgName: elem.img_src,
-        sol: elem.sol,
-        cameraShort: elem.camera.name,
-        cameraFull: elem.camera.full_name,
-        roverName: elem.rover.name,
-        roverStatus: elem.rover.status,
-      };
-      tableInfo.push(tempDay);
-      return true;
-    });
-    this.setState({ table: tableInfo });
-    console.table(this.state.table);
-  };
 
   getTestTable = () => {
     try {
@@ -69,7 +43,7 @@ class ApiGetter extends Component {
   render() {
     return (
       <div className="api-getter">
-        <button className="test-button" onClick={() => this.getInfoFromAPI()}>
+        <button className="test-button" onClick={async () => this.setState(await getInfoFromAPI())}>
           Test
         </button>
         {this.getTestTable()}
